// WebApplicationPlateforme/ClientApp/src/app/shared/Models/MediaCenter/ImpressionDesign/design-impression.model.ts
export class DesignImpression {

  id: number;
  nomdesign: string;
  etat: string;
  iddir: string;
  nomdir: string;
  datedir: string;
  transfertA: string;
  transfertAdmin: string;
  transfertEmployee: string;
  datetransfertAdmin: string;
  dattransfertEmployee: string;
  idadminEmp: string;
  nomadminEmp: string;
  dateadminEmp: string;
  etatidadminEmp: string;
  dateDebadminEmp: string;
  dateFinadminEmp: string;
  dateRecepadminEmp: string;
  dateDoneadminEmp: string;
  transfertdminEmp: string;
  idEmp: string;
  nomEmp: string;
  dateEmp: string;
  etatEmp: string;
  tranfertEmp: string;
  dateDebutEmp: string;
  dateFinEmp: string;
  dateReceptiondateEmp: string;
  dateTransfertEmp: string;
  tavailfait: string;
  adminid: string;
  adminom: string;
  admietat: string;
  admidate: string;
  etabnom: string;
  etabid: string;
  etabetat: string;
  etabdate: string;
  attribut1: string;
  attribut2: string;
  attribut3: string;
  attribut4: string;
  attribut5: string;
  attribut6: string;
  dateenreg: string;
  userNameCreator: string;
  idUserCreator: string;

}